refactor(editor): Isolate n8n Connect section logic from generic node creator components
